more refactoring done
